Basic web server created
The ability to establish a connection with the server has been created. fs lib imported, and used to read html pages that will be shown to the client depending on success/error. Very basic logic has been implemented to control the whether the request is for the home page, otherwise, they're redirected to a 404.html page.